requirements.txt
# django
django==3.2
django-environ==0.9.0
django-filter==22.1

# DRF
djangorestframework==3.13.1
dj-rest-auth[with_social]
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.23.1
drf-spectacular-sidecar==2022.8.1
drf-psq==1.1.0

# DB
psycopg2-binary==2.9.3
pillow==9.2.0
django-phonenumber-field==6.4.0
phonenumbers==8.12.54
Faker==14.2.0

# Celery
celery==5.2.7
redis==4.3.4

# Testing
pytest==7.1.3
pytest-django==4.5.2
model-bakery==1.7.0

